[Framework] Move publically used private headers in to interface dir
