+++ FWCore/Framework/interface/streamTransitionAsync.h
@@ -19,7 +19,7 @@
 //
 
 #include "FWCore/Framework/interface/Schedule.h"
-#include "FWCore/Framework/src/SubProcess.h"
+#include "FWCore/Framework/interface/SubProcess.h"
 #include "FWCore/Framework/interface/TransitionInfoTypes.h"
 #include "FWCore/Concurrency/interface/WaitingTask.h"
 #include "FWCore/Concurrency/interface/WaitingTaskHolder.h"
